XER10-117: Adding XER10 specific flags in Utopia
Reason for change: Adding XER10 specific flags in Utopia
Test Procedure: Build and check
Risks: Low
Priority: P1

(cherry picked from commit 375b29253d64d33e7054270bf957df6f661141a8)

+++ source/scripts/init/service.d/service_multinet/handle_gre.sh
@@ -265,12 +265,12 @@ create_tunnel () {
         fi
     fi
     ifconfig $2 up
-    if [ ! -f /tmp/.gre_flowmanager_enable ] && [ "$BOX_TYPE" != "VNTXER5" ]
+    if [ ! -f /tmp/.gre_flowmanager_enable ] && [ "$BOX_TYPE" != "VNTXER5" ] && [ "$BOX_TYPE" != "SCER11BEL" ]
     then
           echo addif $2 wan > /proc/driver/flowmgr/cmd
 	  touch /tmp/.gre_flowmanager_enable
     fi
-    if [ "$BOX_TYPE" = "VNTXER5" ]; then
+    if [ "$BOX_TYPE" = "VNTXER5" ] || [ "$BOX_TYPE" = "SCER11BEL" ]; then
 	    INST=`psmcli getallinst dmsb.l2net. | tr -d \n | sed ' s/./& /g'`
 	    for i in $INST; do
 		    GRE=`psmcli get dmsb.l2net.$i.Members.Gre | grep $2`
@@ -303,7 +303,7 @@ destroy_tunnel () {
       touch "/tmp/destroy_tunnel_lock"
    fi
     echo "Destroying tunnel... remote"
-    if [ "$BOX_TYPE" = "VNTXER5" ]; then
+    if [ "$BOX_TYPE" = "VNTXER5" ] || [ "$BOX_TYPE" = "SCER11BEL" ]; then
 	    INST=`psmcli getallinst dmsb.l2net. | tr -d \n | sed ' s/./& /g'`
 	    for i in $INST; do
 		    STAT=`sysevent get gre_"$i"_inst`
@@ -505,7 +505,7 @@ update_bridge_config () {
             continue
         fi
 
-    if [ "$BOX_TYPE" != "VNTXER5" ]; then
+    if [ "$BOX_TYPE" != "VNTXER5" ] && [ "$BOX_TYPE" != "SCER11BEL" ]; then
         br_snoop_rule="`sysevent setunique GeneralPurposeFirewallRule " -A FORWARD -o $br -p udp --dport=67:68 -j NFQUEUE --queue-bypass --queue-num $[$br]"`"
         sysevent set gre_${inst}_${br}_snoop_rule "$br_snoop_rule"
     fi
@@ -902,7 +902,7 @@ case "$1" in
             if [ "$BOX_TYPE" = "XF3" ] ; then
             sleep 5
             fi
-	    if [ "$BOX_TYPE" != "VNTXER5" ]; then
+	    if [ "$BOX_TYPE" != "VNTXER5" ] && [ "$BOX_TYPE" != "SCER11BEL" ]; then
 		    arpFWrule=`sysevent setunique GeneralPurposeFirewallRule " -I OUTPUT -o $WAN_IF -p icmp --icmp-type 3 -j NFQUEUE --queue-bypass --queue-num $ARP_NFQUEUE"`
 		    sysevent set ${inst}_arp_queue_rule "$arpFWrule" > /dev/null
 	    fi
